[REFACTOR] Moved get_media_path into File_Tracking
The file object should have full control of the paths and objects so we
can control locations from there.
