+++ api/user/models.py
@@ -123,7 +123,6 @@ class User(UserMixin, db.Document):
         return get_response_error_formatted(403, {'error_msg': 'User is not active!'})
 
     def delete_media(self):
-        from api.media.routes import get_media_path
         from api.media.models import File_Tracking
 
         print(" FULL USER CLEAN UP - REMOVE FILES AND DATABASE ENTRIES ")
@@ -132,7 +131,7 @@ class User(UserMixin, db.Document):
         File_Tracking.objects(username=self.username).delete()
 
         # Delete the entire folder path
-        full_path = get_media_path() + self.username + "/"
+        full_path = File_Tracking.get_media_path() + self.username + "/"
 
         if os.path.exists(full_path):
             shutil.rmtree(full_path)
